Split large emit_wasm files into focused modules
