+++ src/codegen/emit_wasm/calls.rs
@@ -1,8 +1,8 @@
 //! Function call emission — emit_call and related helpers.
 
-use crate::ir::{CallTarget, IrExpr, IrStringPart};
+use crate::ir::{CallTarget, IrExpr};
 use crate::types::Ty;
-use wasm_encoder::{Instruction, ValType};
+use wasm_encoder::ValType;
 
 use super::FuncCompiler;
 use super::values;
@@ -165,839 +165,13 @@ impl FuncCompiler<'_> {
                             self.emit_stub_call(args);
                         }
                     }
-                    ("list", "map") => {
-                        self.emit_list_map(&args[0], &args[1], _ret_ty);
-                    }
-                    ("list", "enumerate") => {
-                        // enumerate(list) → list of (index, element) tuples
-                        // Each tuple is heap-allocated: [Int(i64), element]
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let tuple_size = 8 + elem_size; // Int(8) + elem
-
-                        let s = self.match_i32_base + self.match_depth;
-                        let len_local = s;
-                        let idx_local = s + 1;
-
-                        // Store src → mem[0]
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // len
-                            i32_const(0);
-                            i32_load(0);
-                            i32_load(0);
-                            local_set(len_local);
-                            // Alloc dst: [len] + len * ptr_size(4)
-                            i32_const(8);
-                            i32_const(4);
-                            local_get(len_local);
-                            i32_const(4); // each entry is a tuple ptr (i32)
-                            i32_mul;
-                            i32_add;
-                            call(self.emitter.rt.alloc);
-                            i32_store(0);
-                            // dst = mem[8]
-                            // Store len in dst
-                            i32_const(8);
-                            i32_load(0);
-                            local_get(len_local);
-                            i32_store(0);
-                            // Loop: create tuples
-                            i32_const(0);
-                            local_set(idx_local);
-                            block_empty;
-                            loop_empty;
-                        });
-                        let saved = self.depth;
-                        self.depth += 2;
-
-                        wasm!(self.func, {
-                            local_get(idx_local);
-                            local_get(len_local);
-                            i32_ge_u;
-                            br_if(1);
-                            // Alloc tuple: [index:i64][element]
-                            i32_const(tuple_size as i32);
-                            call(self.emitter.rt.alloc);
-                            // tuple_ptr on stack. Store to mem[12]
-                            // (original code had some stack manipulation; the final approach: drop and re-alloc)
-                            drop;
-                            // Re-alloc tuple
-                            i32_const(12);
-                            i32_const(tuple_size as i32);
-                            call(self.emitter.rt.alloc);
-                            i32_store(0); // mem[12] = tuple_ptr
-                            // tuple.index = idx (as i64)
-                            i32_const(12);
-                            i32_load(0); // tuple_ptr
-                            local_get(idx_local);
-                            i64_extend_i32_u;
-                            i64_store(0);
-                            // tuple.element = src[idx]
-                            i32_const(12);
-                            i32_load(0); // tuple_ptr
-                            // Load src element
-                            i32_const(0);
-                            i32_load(0); // src_ptr
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0);
-                        self.emit_store_at(&elem_ty, 8); // store at tuple offset 8
-
-                        wasm!(self.func, {
-                            // dst[idx] = tuple_ptr
-                            i32_const(8);
-                            i32_load(0); // dst_ptr
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(4); // tuple ptrs are i32
-                            i32_mul;
-                            i32_add;
-                            i32_const(12);
-                            i32_load(0); // tuple_ptr
-                            i32_store(0);
-                            // idx++
-                            local_get(idx_local);
-                            i32_const(1);
-                            i32_add;
-                            local_set(idx_local);
-                            br(0);
-                        });
-
-                        self.depth = saved;
-                        wasm!(self.func, {
-                            end;
-                            end;
-                            // Return dst
-                            i32_const(8);
-                            i32_load(0);
-                        });
-                    }
-                    ("list", "get") => {
-                        // list.get(list, index) → Option[T]
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-
-                        // mem[0]=list, mem[4]=idx(i32)
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            i32_const(4);
-                        });
-                        self.emit_expr(&args[1]);
-                        if matches!(&args[1].ty, Ty::Int) {
-                            wasm!(self.func, { i32_wrap_i64; });
-                        }
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // bounds: idx >= len → none(0)
-                            i32_const(4);
-                            i32_load(0); // idx
-                            i32_const(0);
-                            i32_load(0); // list
-                            i32_load(0); // len
-                            i32_ge_u;
-                            if_i32;
-                            i32_const(0); // none
-                            else_;
-                            // alloc → mem[8]
-                            i32_const(8);
-                            i32_const(elem_size as i32);
-                            call(self.emitter.rt.alloc);
-                            i32_store(0);
-                            // dst=mem[8], src=list+4+idx*elem_size
-                            i32_const(8);
-                            i32_load(0); // dst
-                            i32_const(0);
-                            i32_load(0); // list
-                            i32_const(4);
-                            i32_add;
-                            i32_const(4);
-                            i32_load(0); // idx
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0); // load elem
-                        self.emit_store_at(&elem_ty, 0); // store at dst
-                        wasm!(self.func, {
-                            i32_const(8);
-                            i32_load(0); // return ptr
-                            end;
-                        });
-                    }
-                    ("list", "filter") => {
-                        // filter(list, fn) → new list with matching elements
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let s = self.match_i32_base + self.match_depth;
-                        let len_local = s;
-                        let idx_local = s + 1;
-
-                        // mem[0]=src, mem[4]=fn, mem[8]=dst, mem[12]=out_idx
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            i32_const(4);
-                        });
-                        self.emit_expr(&args[1]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // len
-                            i32_const(0);
-                            i32_load(0);
-                            i32_load(0);
-                            local_set(len_local);
-                            // alloc dst (max size = 4 + len * elem_size) → mem[8]
-                            i32_const(8);
-                            i32_const(4);
-                            local_get(len_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                            call(self.emitter.rt.alloc);
-                            i32_store(0);
-                            // out_idx = 0 → mem[12]
-                            i32_const(12);
-                            i32_const(0);
-                            i32_store(0);
-                            // idx = 0
-                            i32_const(0);
-                            local_set(idx_local);
-                            // Loop
-                            block_empty;
-                            loop_empty;
-                        });
-                        let saved = self.depth; self.depth += 2;
-
-                        wasm!(self.func, {
-                            local_get(idx_local);
-                            local_get(len_local);
-                            i32_ge_u;
-                            br_if(1);
-                            // Call predicate: fn(element) → bool (i32)
-                            // Load closure
-                            i32_const(4);
-                            i32_load(0);
-                        });
-                        wasm!(self.func, {
-                            i32_load(4);
-                            // Load element
-                            i32_const(0);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0);
-                        // table_idx
-                        wasm!(self.func, {
-                            i32_const(4);
-                            i32_load(0);
-                            i32_load(0);
-                        });
-                        // call_indirect
-                        if let Ty::Fn { params, ret } = &args[1].ty {
-                            let mut ct = vec![ValType::I32];
-                            for p in params { if let Some(vt) = values::ty_to_valtype(p) { ct.push(vt); } }
-                            let rt = values::ret_type(ret);
-                            let ti = self.emitter.register_type(ct, rt);
-                            wasm!(self.func, { call_indirect(ti, 0); });
-                        }
-                        // If true, copy element to dst
-                        wasm!(self.func, {
-                            if_empty;
-                            // dst[out_idx] = src[idx]
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(12);
-                            i32_load(0);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                            // load src element
-                            i32_const(0);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0);
-                        self.emit_store_at(&elem_ty, 0);
-                        wasm!(self.func, {
-                            // out_idx++
-                            i32_const(12);
-                            i32_const(12);
-                            i32_load(0);
-                            i32_const(1);
-                            i32_add;
-                            i32_store(0);
-                            end; // end if
-                            // idx++
-                            local_get(idx_local);
-                            i32_const(1);
-                            i32_add;
-                            local_set(idx_local);
-                            br(0);
-                        });
-
-                        self.depth = saved;
-                        wasm!(self.func, {
-                            end;
-                            end;
-                            // Set dst.len = out_idx
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(12);
-                            i32_load(0);
-                            i32_store(0);
-                            // Return dst
-                            i32_const(8);
-                            i32_load(0);
-                        });
-                    }
-                    ("list", "fold") => {
-                        // fold(list, init, fn(acc, elem) → acc)
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let s = self.match_i32_base + self.match_depth;
-                        let len_local = s;
-                        let idx_local = s + 1;
-                        // Accumulator local: use i64 for Int/Float, i32 for everything else
-                        let acc_local = match values::ty_to_valtype(&args[1].ty) {
-                            Some(ValType::I64) | Some(ValType::F64) => self.match_i64_base + self.match_depth,
-                            _ => self.match_i32_base + self.match_depth + 2, // after len + idx
-                        };
-
-                        // mem[0]=list, mem[4]=fn
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, { i32_store(0); });
-                        // acc = init
-                        self.emit_expr(&args[1]);
-                        wasm!(self.func, {
-                            local_set(acc_local);
-                            i32_const(4);
-                        });
-                        self.emit_expr(&args[2]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // len
-                            i32_const(0);
-                            i32_load(0);
-                            i32_load(0);
-                            local_set(len_local);
-                            i32_const(0);
-                            local_set(idx_local);
-                            block_empty;
-                            loop_empty;
-                        });
-                        let saved = self.depth; self.depth += 2;
-
-                        wasm!(self.func, {
-                            local_get(idx_local);
-                            local_get(len_local);
-                            i32_ge_u;
-                            br_if(1);
-                            // acc = fn(acc, elem)
-                            i32_const(4);
-                            i32_load(0);
-                        });
-                        wasm!(self.func, {
-                            i32_load(4);
-                            local_get(acc_local);
-                            // load elem
-                            i32_const(0);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0);
-                        // table_idx
-                        wasm!(self.func, {
-                            i32_const(4);
-                            i32_load(0);
-                            i32_load(0);
-                        });
-                        if let Ty::Fn { params, ret } = &args[2].ty {
-                            let mut ct = vec![ValType::I32];
-                            for p in params { if let Some(vt) = values::ty_to_valtype(p) { ct.push(vt); } }
-                            let rt = values::ret_type(ret);
-                            let ti = self.emitter.register_type(ct, rt);
-                            wasm!(self.func, { call_indirect(ti, 0); });
-                        }
-                        wasm!(self.func, {
-                            local_set(acc_local);
-                            local_get(idx_local);
-                            i32_const(1);
-                            i32_add;
-                            local_set(idx_local);
-                            br(0);
-                        });
-
-                        self.depth = saved;
-                        wasm!(self.func, {
-                            end;
-                            end;
-                            local_get(acc_local);
-                        });
-                    }
-                    ("list", "reverse") => {
-                        // reverse(list) → new list with elements in reverse order
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let s = self.match_i32_base + self.match_depth;
-                        let len_local = s;
-                        let idx_local = s + 1;
-
-                        // mem[0]=src
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // len
-                            i32_const(0);
-                            i32_load(0);
-                            i32_load(0);
-                            local_set(len_local);
-                            // alloc dst → mem[4]
-                            i32_const(4);
-                            i32_const(4);
-                            local_get(len_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                            call(self.emitter.rt.alloc);
-                            i32_store(0);
-                            // dst.len = len
-                            i32_const(4);
-                            i32_load(0);
-                            local_get(len_local);
-                            i32_store(0);
-                            // Loop: dst[len-1-i] = src[i]
-                            i32_const(0);
-                            local_set(idx_local);
-                            block_empty;
-                            loop_empty;
-                        });
-                        let saved = self.depth; self.depth += 2;
-
-                        wasm!(self.func, {
-                            local_get(idx_local);
-                            local_get(len_local);
-                            i32_ge_u;
-                            br_if(1);
-                            // dst addr: dst + 4 + (len-1-i) * elem_size
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            local_get(len_local);
-                            i32_const(1);
-                            i32_sub;
-                            local_get(idx_local);
-                            i32_sub;
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                            // src elem: src + 4 + i * elem_size
-                            i32_const(0);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            local_get(idx_local);
-                            i32_const(elem_size as i32);
-                            i32_mul;
-                            i32_add;
-                        });
-                        self.emit_load_at(&elem_ty, 0);
-                        self.emit_store_at(&elem_ty, 0);
-
-                        wasm!(self.func, {
-                            local_get(idx_local);
-                            i32_const(1);
-                            i32_add;
-                            local_set(idx_local);
-                            br(0);
-                        });
-
-                        self.depth = saved;
-                        wasm!(self.func, {
-                            end;
-                            end;
-                            i32_const(4);
-                            i32_load(0);
-                        });
-                    }
-                    ("list", "sort") => {
-                        // Bubble sort on a copy. Int(i64) only.
-                        // mem[0]=src, alloc copy → mem[4], len → local s, i → local s+1
-                        let s = self.match_i32_base + self.match_depth;
-
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, {
-                            i32_store(0);
-                            // len → local s
-                            i32_const(0);
-                            i32_load(0);
-                            i32_load(0);
-                            local_set(s);
-                            // total_bytes = 4 + len * 8
-                            // alloc → mem[4]
-                            i32_const(4);
-                            i32_const(4);
-                            local_get(s);
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            call(self.emitter.rt.alloc);
-                            i32_store(0);
-                            // Store total_bytes → mem[8]
-                            i32_const(8);
-                            i32_const(4);
-                            local_get(s);
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i32_store(0); // mem[8] = total
-                            // Byte copy loop: i=0..total
-                            i32_const(0);
-                            local_set(s + 1); // i=0
-                            block_empty;
-                            loop_empty;
-                        });
-                        self.depth += 2;
-                        wasm!(self.func, {
-                            local_get(s + 1);
-                            i32_const(8);
-                            i32_load(0); // total
-                            i32_ge_u;
-                            br_if(1);
-                            // dst[i] = src[i] (src=mem[0], dst=mem[4])
-                            i32_const(4);
-                            i32_load(0); // dst
-                            local_get(s + 1);
-                            i32_add;
-                            i32_const(0);
-                            i32_load(0); // src
-                            local_get(s + 1);
-                            i32_add;
-                            i32_load8_u(0);
-                            i32_store8(0);
-                            local_get(s + 1);
-                            i32_const(1);
-                            i32_add;
-                            local_set(s + 1);
-                            br(0);
-                        });
-                        self.depth -= 2;
-                        wasm!(self.func, {
-                            end;
-                            end;
-                            // Bubble sort: outer i=0..len-1, inner j=0..len-1-i
-                            // mem[8]=j (reuse), mem[12]=tmp(i64)
-                            i32_const(0);
-                            local_set(s + 1); // i=0
-                            block_empty;
-                            loop_empty;
-                        });
-                        self.depth += 2;
-                        wasm!(self.func, {
-                            local_get(s + 1);
-                            local_get(s);
-                            i32_const(1);
-                            i32_sub;
-                            i32_ge_u;
-                            br_if(1);
-                            // Inner: j=0
-                            i32_const(8);
-                            i32_const(0);
-                            i32_store(0);
-                            block_empty;
-                            loop_empty;
-                        });
-                        self.depth += 2;
-                        wasm!(self.func, {
-                            i32_const(8);
-                            i32_load(0); // j
-                            local_get(s);
-                            i32_const(1);
-                            i32_sub;
-                            local_get(s + 1);
-                            i32_sub;
-                            i32_ge_u;
-                            br_if(1);
-                            // Compare dst[j] > dst[j+1]
-                            // addr_j = dst + 4 + j*8
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i64_load(0);
-                            // dst[j+1]
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(1);
-                            i32_add;
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i64_load(0);
-                            i64_gt_s;
-                            if_empty;
-                            // Swap: tmp=dst[j], dst[j]=dst[j+1], dst[j+1]=tmp
-                            // tmp → mem[12..20]
-                            i32_const(12);
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i64_load(0);
-                            i64_store(0);
-                            // dst[j] = dst[j+1]
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            // value = dst[j+1]
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(1);
-                            i32_add;
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i64_load(0);
-                            i64_store(0);
-                            // dst[j+1] = tmp(mem[12])
-                            i32_const(4);
-                            i32_load(0);
-                            i32_const(4);
-                            i32_add;
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(1);
-                            i32_add;
-                            i32_const(8);
-                            i32_mul;
-                            i32_add;
-                            i32_const(12);
-                            i64_load(0);
-                            i64_store(0);
-                            end; // end if swap
-                            // j++
-                            i32_const(8);
-                            i32_const(8);
-                            i32_load(0);
-                            i32_const(1);
-                            i32_add;
-                            i32_store(0);
-                            br(0);
-                        });
-                        self.depth -= 2;
-                        wasm!(self.func, {
-                            end; // end inner loop
-                            end; // end inner block
-                            // i++
-                            local_get(s + 1);
-                            i32_const(1);
-                            i32_add;
-                            local_set(s + 1);
-                            br(0);
-                        });
-                        self.depth -= 2;
-                        wasm!(self.func, {
-                            end; // end outer loop
-                            end; // end outer block
-                            // Return dst
-                            i32_const(4);
-                            i32_load(0);
-                        });
-                    }
-                    ("list", "get") => {
-                        // list.get(list, index) -> Option[T]
-                        // Returns some(elem) if in bounds, none otherwise
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let s = self.match_i32_base + self.match_depth;
-                        // Store list → mem[0], index → mem[4]
-                        wasm!(self.func, { i32_const(0); });
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, { i32_store(0); });
-                        self.emit_expr(&args[1]);
-                        wasm!(self.func, {
-                            i32_wrap_i64; // index to i32
-                            local_set(s);
-                            // Check bounds: index < 0 || index >= len → none (null ptr = 0)
-                            local_get(s);
-                            i32_const(0);
-                            i32_lt_u;
-                            local_get(s);
-                            i32_const(0);
-                            i32_load(0); // list
-                            i32_load(0); // len
-                            i32_ge_u;
-                            i32_or;
-                            if_i32;
-                            i32_const(0); // none
-                            else_;
-                            // some: alloc elem_size, copy value
-                            i32_const(elem_size as i32);
-                            call(self.emitter.rt.alloc);
-                            local_set(s);
-                            // Copy: src = list_ptr + 4 + index * elem_size
-                            local_get(s);
-                            i32_const(0);
-                            i32_load(0); // list
-                            i32_const(4);
-                            i32_add;
-                        });
-                        wasm!(self.func, {
-                            local_get(s);
-                        });
-                        // Hmm, this is getting complex. Use simpler approach:
-                        // Actually, Option[T] is represented as: none=null(0), some=ptr to T.
-                        // For list.get, return ptr to element directly (no copy needed for i32 types)
-                        // But for i64/f64 we need to allocate.
-                        // Simplify: allocate and copy for all types.
-                        wasm!(self.func, {
-                            // Scratch already has alloc'd ptr. Need to write element there.
-                            // Reset: use a different approach
-                            drop; drop; // drop the half-built stack
-                        });
-                        // TODO: implement properly. For now, return none (0) always.
-                        wasm!(self.func, { i32_const(0); });
-                    }
-                    ("list", "contains") => {
-                        // list.contains(list, elem) -> Bool (i32)
-                        // Uses only scratch locals — no mem[] to avoid nested call conflicts
-                        let elem_ty = if let Ty::Applied(_, a) = &args[0].ty {
-                            a.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        let elem_size = values::byte_size(&elem_ty);
-                        let s = self.match_i32_base + self.match_depth;
-                        // s=list_ptr, s+1=idx, s+2=result
-                        self.emit_expr(&args[0]);
-                        wasm!(self.func, { local_set(s); });
-                        // Save target to i64 scratch or i32 scratch depending on type
-                        match values::ty_to_valtype(&elem_ty) {
-                            Some(ValType::I64) => {
-                                let t = self.match_i64_base + self.match_depth;
-                                self.emit_expr(&args[1]);
-                                wasm!(self.func, {
-                                    local_set(t); // target in i64 local
-                                    i32_const(0); local_set(s + 1); // idx
-                                    i32_const(0); local_set(s + 2); // result = false
-                                    block_empty; loop_empty;
-                                      local_get(s + 1); local_get(s); i32_load(0); i32_ge_u; br_if(1);
-                                      local_get(s); i32_const(4); i32_add;
-                                      local_get(s + 1); i32_const(elem_size as i32); i32_mul; i32_add;
-                                      i64_load(0);
-                                      local_get(t); i64_eq;
-                                      if_empty;
-                                        i32_const(1); local_set(s + 2); br(2);
-                                      end;
-                                      local_get(s + 1); i32_const(1); i32_add; local_set(s + 1);
-                                      br(0);
-                                    end; end;
-                                    local_get(s + 2);
-                                });
-                            }
-                            _ => {
-                                // i32 types: String, Option, etc.
-                                self.emit_expr(&args[1]);
-                                wasm!(self.func, {
-                                    local_set(s + 3); // target in i32 local
-                                    i32_const(0); local_set(s + 1);
-                                    i32_const(0); local_set(s + 2);
-                                    block_empty; loop_empty;
-                                      local_get(s + 1); local_get(s); i32_load(0); i32_ge_u; br_if(1);
-                                      local_get(s); i32_const(4); i32_add;
-                                      local_get(s + 1); i32_const(elem_size as i32); i32_mul; i32_add;
-                                      i32_load(0);
-                                      local_get(s + 3);
-                                });
-                                match &elem_ty {
-                                    Ty::String => { wasm!(self.func, { call(self.emitter.rt.string.eq); }); }
-                                    _ => { wasm!(self.func, { i32_eq; }); }
-                                }
-                                wasm!(self.func, {
-                                      if_empty;
-                                        i32_const(1); local_set(s + 2); br(2);
-                                      end;
-                                      local_get(s + 1); i32_const(1); i32_add; local_set(s + 1);
-                                      br(0);
-                                    end; end;
-                                    local_get(s + 2);
-                                });
-                            }
+                    _ if module == "list" => {
+                        if !self.emit_list_call(func, args) {
+                            self.emit_stub_call(args);
                         }
                     }
                     _ if module == "map" => {
                         if !self.emit_map_call(func, args) {
-                            self.emit_stub_call(args);
-                        }
-                    }
-                    _ if module == "list" => {
-                        if !self.emit_list_call(func, args) {
                             self.emit_stub_call(args);
                         }
                     }
@@ -1293,236 +467,6 @@ impl FuncCompiler<'_> {
         }
     }
 
-    /// Emit a FnRef as a closure: allocate [wrapper_table_idx, 0] on heap.
-    pub(super) fn emit_fn_ref_closure(&mut self, name: &str) {
-        if let Some(&wrapper_table_idx) = self.emitter.fn_ref_wrappers.get(name) {
-            // Allocate closure: [table_idx: i32][env_ptr: i32] = 8 bytes
-            let scratch = self.match_i32_base + self.match_depth;
-            wasm!(self.func, {
-                i32_const(8);
-                call(self.emitter.rt.alloc);
-                local_set(scratch);
-                // Store table_idx
-                local_get(scratch);
-                i32_const(wrapper_table_idx as i32);
-                i32_store(0);
-                // Store env_ptr = 0
-                local_get(scratch);
-                i32_const(0);
-                i32_store(4);
-                // Return closure ptr
-                local_get(scratch);
-            });
-        } else {
-            eprintln!("WARNING: FnRef wrapper not found for '{}', using direct table entry", name);
-            wasm!(self.func, { unreachable; });
-        }
-    }
-
-    /// Emit a lambda as a closure: allocate env + closure on heap.
-    pub(super) fn emit_lambda_closure(&mut self, _params: &[(crate::ir::VarId, Ty)], _body: &IrExpr) {
-        // Match lambda by param VarIds (deterministic, order-independent of scan)
-        let param_key: Vec<u32> = _params.iter().map(|(vid, _)| vid.0).collect();
-        let lambda_idx = self.emitter.lambdas.iter().enumerate()
-            .find(|(_, info)| info.param_ids == param_key)
-            .map(|(i, _)| i);
-
-        let lambda_idx = match lambda_idx {
-            Some(i) => i,
-            None => {
-                // Fallback to counter (for compatibility)
-                let idx = self.emitter.lambda_counter.get();
-                self.emitter.lambda_counter.set(idx + 1);
-                if idx >= self.emitter.lambdas.len() {
-                    wasm!(self.func, { unreachable; });
-                    return;
-                }
-                idx
-            }
-        };
-
-        let table_idx = self.emitter.lambdas[lambda_idx].table_idx;
-        let captures = self.emitter.lambdas[lambda_idx].captures.clone();
-
-        let scratch = self.match_i32_base + self.match_depth;
-
-        if captures.is_empty() {
-            // No captures: allocate closure [table_idx, 0]
-            wasm!(self.func, {
-                i32_const(8);
-                call(self.emitter.rt.alloc);
-                local_set(scratch);
-                local_get(scratch);
-                i32_const(table_idx as i32);
-                i32_store(0);
-                local_get(scratch);
-                i32_const(0);
-                i32_store(4);
-                local_get(scratch);
-            });
-        } else {
-            // Allocate env: each capture gets 8 bytes (padded for alignment)
-            let env_size = (captures.len() as u32) * 8;
-            let env_scratch = scratch; // reuse for env_ptr
-            wasm!(self.func, {
-                i32_const(env_size as i32);
-                call(self.emitter.rt.alloc);
-                local_set(env_scratch);
-            });
-
-            // Store each captured variable into env
-            for (ci, (vid, ty)) in captures.iter().enumerate() {
-                let offset = (ci as u32) * 8;
-                let is_cell = self.emitter.mutable_captures.contains(&vid.0);
-                wasm!(self.func, { local_get(env_scratch); });
-                if let Some(&local_idx) = self.var_map.get(&vid.0) {
-                    wasm!(self.func, { local_get(local_idx); });
-                    if is_cell {
-                        // Mutable capture: local is cell ptr (i32), store as i32
-                        wasm!(self.func, { i32_store(offset); });
-                    } else {
-                        self.emit_store_at(ty, offset);
-                    }
-                } else {
-                    match values::ty_to_valtype(ty) {
-                        Some(ValType::I64) => { wasm!(self.func, { i64_const(0); }); }
-                        Some(ValType::F64) => { wasm!(self.func, { f64_const(0.0); }); }
-                        _ => { wasm!(self.func, { i32_const(0); }); }
-                    }
-                    self.emit_store_at(ty, offset);
-                }
-            }
-
-            // Allocate closure: [table_idx, env_ptr]
-            let closure_scratch = scratch + 1; // second i32 scratch slot
-            wasm!(self.func, {
-                i32_const(8);
-                call(self.emitter.rt.alloc);
-                local_set(closure_scratch);
-                local_get(closure_scratch);
-                i32_const(table_idx as i32);
-                i32_store(0);
-                local_get(closure_scratch);
-                local_get(env_scratch);
-                i32_store(4);
-                local_get(closure_scratch);
-            });
-        }
-    }
-
-    /// ASCII case conversion. Expects string ptr on stack. Returns new string ptr.
-    fn emit_str_case_convert(&mut self, is_upper: bool) {
-        // String ptr is on stack. Store to mem[0] via scratch.
-        let scratch = self.match_i32_base + self.match_depth;
-        wasm!(self.func, {
-            local_set(scratch);
-            i32_const(0);
-            local_get(scratch);
-            i32_store(0);
-            // Alloc dst with same len → mem[4]
-            i32_const(4);
-            i32_const(4);
-            i32_const(0);
-            i32_load(0);
-            i32_load(0);
-            i32_add;
-            call(self.emitter.rt.alloc);
-            i32_store(0);
-            // Store len in dst
-            i32_const(4);
-            i32_load(0);
-            i32_const(0);
-            i32_load(0);
-            i32_load(0);
-            i32_store(0);
-        });
-        // Loop: convert each byte
-        let s = self.match_i32_base + self.match_depth;
-        wasm!(self.func, {
-            i32_const(0);
-            local_set(s);
-            block_empty;
-            loop_empty;
-        });
-        let saved = self.depth; self.depth += 2;
-        wasm!(self.func, {
-            local_get(s);
-            i32_const(0);
-            i32_load(0);
-            i32_load(0);
-            i32_ge_u;
-            br_if(1);
-            // dst addr
-            i32_const(4);
-            i32_load(0);
-            i32_const(4);
-            i32_add;
-            local_get(s);
-            i32_add;
-            // src byte
-            i32_const(0);
-            i32_load(0);
-            i32_const(4);
-            i32_add;
-            local_get(s);
-            i32_add;
-            i32_load8_u(0);
-            // Convert
-            local_set(s + 1);
-        });
-        if is_upper {
-            wasm!(self.func, {
-                local_get(s + 1);
-                i32_const(97);
-                i32_ge_u;
-                local_get(s + 1);
-                i32_const(122);
-                i32_le_u;
-                i32_and;
-                if_i32;
-                local_get(s + 1);
-                i32_const(32);
-                i32_sub;
-                else_;
-                local_get(s + 1);
-                end;
-            });
-        } else {
-            wasm!(self.func, {
-                local_get(s + 1);
-                i32_const(65);
-                i32_ge_u;
-                local_get(s + 1);
-                i32_const(90);
-                i32_le_u;
-                i32_and;
-                if_i32;
-                local_get(s + 1);
-                i32_const(32);
-                i32_add;
-                else_;
-                local_get(s + 1);
-                end;
-            });
-        }
-        wasm!(self.func, {
-            i32_store8(0);
-            local_get(s);
-            i32_const(1);
-            i32_add;
-            local_set(s);
-            br(0);
-        });
-        self.depth = saved;
-        wasm!(self.func, {
-            end;
-            end;
-            // Return dst
-            i32_const(4);
-            i32_load(0);
-        });
-    }
-
     /// Emit a stub for an unimplemented call: evaluate args (for side effects), drop values, unreachable.
     pub(super) fn emit_stub_call_logged(&mut self, args: &[IrExpr], context: &str) {
         eprintln!("[STUB] {}", context);
@@ -1639,200 +583,4 @@ impl FuncCompiler<'_> {
             end;
         });
     }
-
-    /// Concatenate two strings on the heap via __concat_str runtime.
-    pub(super) fn emit_concat_str(&mut self, left: &IrExpr, right: &IrExpr) {
-        self.emit_expr(left);
-        self.emit_expr(right);
-        wasm!(self.func, { call(self.emitter.rt.concat_str); });
-    }
-
-    /// String interpolation: convert each part to string, then concat.
-    pub(super) fn emit_string_interp(&mut self, parts: &[IrStringPart]) {
-        if parts.is_empty() {
-            let empty = self.emitter.intern_string("");
-            wasm!(self.func, { i32_const(empty as i32); });
-            return;
-        }
-
-        // Emit first part as a string
-        self.emit_string_part(&parts[0]);
-
-        // For each subsequent part: emit it, then concat with accumulator
-        for part in &parts[1..] {
-            self.emit_string_part(part);
-            wasm!(self.func, { call(self.emitter.rt.concat_str); });
-        }
-    }
-
-    /// Emit a single string interpolation part as a string (i32 pointer).
-    pub(super) fn emit_string_part(&mut self, part: &IrStringPart) {
-        match part {
-            IrStringPart::Lit { value } => {
-                let offset = self.emitter.intern_string(value);
-                wasm!(self.func, { i32_const(offset as i32); });
-            }
-            IrStringPart::Expr { expr } => {
-                match &expr.ty {
-                    Ty::String => self.emit_expr(expr),
-                    Ty::Int => {
-                        self.emit_expr(expr);
-                        wasm!(self.func, { call(self.emitter.rt.int_to_string); });
-                    }
-                    Ty::Bool => {
-                        self.emit_expr(expr);
-                        let t = self.emitter.intern_string("true");
-                        let f = self.emitter.intern_string("false");
-                        wasm!(self.func, {
-                            if_i32;
-                            i32_const(t as i32);
-                            else_;
-                            i32_const(f as i32);
-                            end;
-                        });
-                    }
-                    Ty::Float => {
-                        self.emit_expr(expr);
-                        wasm!(self.func, {
-                            call(self.emitter.rt.float_to_string);
-                        });
-                    }
-                    _ => {
-                        // Fallback: emit the expression (already a string pointer or unsupported)
-                        self.emit_expr(expr);
-                    }
-                }
-            }
-        }
-    }
-
-    /// Emit list.map(list, fn) → new list.
-    /// Uses memory scratch [0..12] for src_ptr/fn_ptr/dst_ptr.
-    /// Key insight: compute dst address BEFORE call_indirect so result goes
-    /// directly onto the stack in the right position for store.
-    fn emit_list_map(&mut self, list_arg: &IrExpr, fn_arg: &IrExpr, ret_ty: &Ty) {
-        let in_elem_ty = if let Ty::Applied(_, args) = &list_arg.ty {
-            args.first().cloned().unwrap_or(Ty::Int)
-        } else { Ty::Int };
-        let out_elem_ty = if let Ty::Applied(_, args) = ret_ty {
-            args.first().cloned().unwrap_or(Ty::Int)
-        } else { Ty::Int };
-        let in_size = values::byte_size(&in_elem_ty);
-        let out_size = values::byte_size(&out_elem_ty);
-
-        let s = self.match_i32_base + self.match_depth;
-        let len_local = s;
-        let idx_local = s + 1;
-        let src_local = s + 2;
-        let closure_local = s + 3;
-
-        // Store src_ptr and closure in scratch locals (not mem[]) to survive nested calls
-        self.emit_expr(list_arg);
-        wasm!(self.func, { local_set(src_local); });
-        self.emit_expr(fn_arg);
-        wasm!(self.func, {
-            local_set(closure_local);
-            local_get(src_local);
-            i32_load(0);
-            local_set(len_local);
-            // Alloc dst
-            i32_const(4);
-            local_get(len_local);
-            i32_const(out_size as i32);
-            i32_mul;
-            i32_add;
-            call(self.emitter.rt.alloc);
-            local_set(s + 4); // dst_local
-        });
-        let dst_local = s + 4;
-        wasm!(self.func, {
-            // Set dst.len
-            local_get(dst_local);
-            local_get(len_local);
-            i32_store(0);
-            // idx = 0
-            i32_const(0);
-            local_set(idx_local);
-            // Loop
-            block_empty;
-            loop_empty;
-        });
-        let saved = self.depth;
-        self.depth += 2;
-
-        wasm!(self.func, {
-            local_get(idx_local);
-            local_get(len_local);
-            i32_ge_u;
-            br_if(1);
-            // dst addr
-            local_get(dst_local);
-            i32_const(4);
-            i32_add;
-            local_get(idx_local);
-            i32_const(out_size as i32);
-            i32_mul;
-            i32_add;
-            // env_ptr from closure
-            local_get(closure_local);
-            i32_load(4);
-            // src element
-            local_get(src_local);
-            i32_const(4);
-            i32_add;
-            local_get(idx_local);
-            i32_const(in_size as i32);
-            i32_mul;
-            i32_add;
-        });
-        self.emit_load_at(&in_elem_ty, 0);
-        // table_idx from closure
-        wasm!(self.func, {
-            local_get(closure_local);
-            i32_load(0);
-        });
-        // Stack: [dst_elem_addr, env_ptr, element, table_idx]
-
-        // call_indirect (env, element) → result
-        // Use fn_arg.ty (Fn type) if available for accurate param/ret types,
-        // otherwise fall back to element types
-        {
-            let ti = if let Ty::Fn { params, ret } = &fn_arg.ty {
-                let mut ct = vec![ValType::I32]; // env
-                for p in params {
-                    if let Some(vt) = values::ty_to_valtype(p) { ct.push(vt); }
-                }
-                let rt = values::ret_type(ret);
-                self.emitter.register_type(ct, rt)
-            } else {
-                let mut ct = vec![ValType::I32];
-                if let Some(vt) = values::ty_to_valtype(&in_elem_ty) { ct.push(vt); }
-                let rt = values::ret_type(&out_elem_ty);
-                self.emitter.register_type(ct, rt)
-            };
-            wasm!(self.func, { call_indirect(ti, 0); });
-        }
-        // Stack: [dst_elem_addr, result]
-
-        // ── Store result at dst addr ──
-        self.emit_store_at(&out_elem_ty, 0);
-        // Stack: []
-
-        // idx++
-        wasm!(self.func, {
-            local_get(idx_local);
-            i32_const(1);
-            i32_add;
-            local_set(idx_local);
-            br(0);
-        });
-
-        self.depth = saved;
-        wasm!(self.func, {
-            end;
-            end;
-            local_get(dst_local);
-        });
-    }
-
 }
