+++ src/codegen/emit_wasm/mod.rs
@@ -25,8 +25,11 @@ mod calls_string;
 mod calls_option;
 mod calls_numeric;
 mod calls_list;
+mod calls_list_closure;
+mod calls_lambda;
 mod calls_map;
 mod calls_set;
+mod equality;
 mod collections;
 mod control;
 pub mod statements;
